Sort: Use anonymous namespace instead of internal
Anonymous namespaces are used to hide the internal
functions from other translation units and are used
for internal implementation functions in other parts
of the library. Use this method in sorting library as
well. A side effect of this caused a rename of the
partition function to partitionImpl because of
ambiguity of the name "partition" with std::partition
caused due to argument dependent lookup.