Remove usage of all anonymous namespaces
1. The use of anonymous namespaces in header files confines names to the
translation unit but not to a file. Therefore there were multiple name
collisions in tree classes. To fix this issue, put all names that need to
be hidden from user in an internal namespace called det (detail) and change
names of the colliding tree classes like TreeNode, TreeVisitor etc. to
more descriptive names like BinaryTreeNode etc.

2. Move tree visitor classes to ad::tree namespace instead of internal
det namespace. This could be useful for ADL incase of implementation of
a generic operator <<, >> for any visitor.

3. Minor formatting changes related to namespace declaration spacing &
comments.