Change det namespaces to private file namespaces
Since the names in det were really private to the file scope, replace
det namespaces with private namespaces with same name as the file. This
makes a nice structure, where a file in <dir1>/<dir2>/<file> will
have the namespace hierarchy dir1::dir2::_file. The underscore indicates
the namespace is private.