Bug: Don't compare tree pointers in CBTVisitor operator==
A default constructed visitor is the end() visitor & it has the
tree pointer as nullptr. If tree pointers are compared in operator==
then it will result in false when idx > tree's size. Options to
fix include checking for idx > tree's size in first() & last() or
don't compare the tree pointers. The latter is implemented for
efficiency.