clang: fix compile error
The forEach_ & transform_ meta algorithms used to take a template
template parameters with 1 & 2 parameters respectively. However this
meant that templates with variadic parameters could not be used to
instantiate these algorithms. Therefore these algorithms are modified
to take a variadic template as their parameter. Note that this was not
an issue in g++, but clang fails to compile the code.